BLD: Generate dynamic files
