--- scipy/version.py
@@ -1,11 +1,0 @@
-
-# THIS FILE IS GENERATED FROM SCIPY SETUP.PY
-short_version = '1.8.0'
-version = '1.8.0'
-full_version = '1.8.0.dev0+1529.461f198'
-git_revision = '461f198'
-commit_count = '1529'
-release = False
-
-if not release:
-    version = full_version
